Update QuickTesting.java
totalTests, totalPass and totalFail are static variables, so we need to reset them in the begin() method.

I added some code in the testingComplete() and testPrintOut() methods do add some colours. I know it works on MacOS, probably will work in any unix like systems (need to test), but I am not sure if it will work on Windows.